Bug fix: display form errors when creating new tenant
in case of form validation errors the page is displayed again but
previous implementation was hiding the error(s) b/c it was creating
a new form instance unconditionally.

With this chenge if a form instance already exists, possibly
containing errors, it will be reused. Otherwise we default to creating
a new form instance. The existing test should report PASS.