Adjust the fallback match string for Kiwi TCMS Enterprise on FastSpring
this only affects old subscriptions which don't have a SKU recorded on
their orders.

We're about to introduce new subscription types which will use the
format kiwitcms-<product>-<qualifier>-subscription and we'd like to have
the fallback trigger on them as well (in case there is ever a problem
with their SKUs).