Add the IPrefixFor field lookup
which reverses the places of left & right operators for the standard
istarts with field. IOW look only for rows where the value of the
specified field acts as a prefix to the string supplied as query
argument!

NOTE: Disable warnings for abstract methods because these methods have
not been implemented in any of the parent classes coming from
django.db.models.lookups either

W0223: Method '__rand__' is abstract in class 'Combinable' but is not overridden in child class 'IPrefixFor' (abstract-method)
W0223: Method '__ror__' is abstract in class 'Combinable' but is not overridden in child class 'IPrefixFor' (abstract-method)
W0223: Method '__rxor__' is abstract in class 'Combinable' but is not overridden in child class 'IPrefixFor' (abstract-method)